Fix form radio field.
Checking a radio input via "checked" attribute from code updates
DOM but somewhy doesn't re-render it in a browser. It looks like
a bug in a browser because DOM was updated that is showed in the
developer console but visually nothing happened.

So I have get it around with flag style classes.